requirements.txt
blynk-library-python==0.2.0
RPi.GPIO==0.6.5
